move files into auth folder for organization
